ocr-service/requirements.txt
fastapi
uvicorn
easyocr
opencv-python-headless
python-multipart
Pillow
numpy
